aaaaaaa

+++ typical90/★2/022.cpp
@@ -1,28 +1,7 @@
 /* include */
-// #include<bits/stdc++.h>
-#include <algorithm>
-#include <bitset>
-#include <cctype>
-#include <cmath>
-#include <cstdio>
-#include <cstdint>
-#include <cstdlib>
-#include <cstring>
-#include <ctime>
-#include <deque>
-#include <functional>
-#include <iomanip>
-#include <iostream>
-#include <list>
-#include <map>
-#include <numeric>
-#include <set>
-#include <sstream>
-#include <stack>
-#include <string>
-#include <utility>
-#include <vector>
+#include<bits/stdc++.h>
 using namespace std;
+// #include<atcoder/all> using namespace atcoder;
 
 /* Macro */
 #define all(container) (container).begin(), (container).end()
@@ -35,23 +14,24 @@ using namespace std;
 /* Type */
 // #define int long long
 #define cauto const auto&
-typedef long long int64;
+typedef long long i64;
 typedef long double ld;
 
 /* Const */
 const double PI = acos(-1.0);
-const int64 MOD = 1000000007;
-const int64 _MOD = 998244353;
+const long long  MOD = 1000000007;
+const long long _MOD = 998244353;
 
-// euclidean algorithm
+
 signed main ()
 {
-  int64 a, b, c;
-  cin >> a >> b >> c;
+  long long a, b, c; cin >> a >> b >> c;
 
-  int64 r = gcd(a, gcd(b, c));
+  long long r = gcd(a, gcd(b, c));
 
   cout << (a / r - 1) + (b / r - 1) + (c / r - 1) << endl;
+
+  // ユークリッドの互除法ンゴねえ
 
   return 0;
 }
